(eval) modified some details about `Evaluate`

+++ src/irgen/eval.rs
@@ -4,17 +4,17 @@ use crate::ast::*;
 
 /// Trait for evaluating constant.
 pub trait Evaluate {
-  fn eval(&self, scopes: &Scopes) -> Option<i64>;
+  fn eval(&self, scopes: &Scopes) -> Option<i32>;
 }
 
 impl Evaluate for Exp {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     self.lor.eval(scopes)
   }
 }
 
 impl Evaluate for LVal {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     let val = scopes.value(&self.id).ok()?;
     if self.indices.is_empty() {
       match val {
@@ -28,7 +28,7 @@ impl Evaluate for LVal {
 }
 
 impl Evaluate for PrimaryExp {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     match self {
       Self::Exp(exp) => exp.eval(scopes),
       Self::LVal(lval) => lval.eval(scopes),
@@ -38,20 +38,20 @@ impl Evaluate for PrimaryExp {
 }
 
 impl Evaluate for UnaryExp {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     match self {
       Self::Primary(primary) => primary.eval(scopes),
       Self::Call(_) => None,
       Self::Unary(op, exp) => exp.eval(scopes).map(|exp| match op {
         UnaryOp::Neg => -exp,
-        UnaryOp::LNot => (exp == 0) as i64,
+        UnaryOp::LNot => (exp == 0) as i32,
       }),
     }
   }
 }
 
 impl Evaluate for MulExp {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     match self {
       Self::Unary(exp) => exp.eval(scopes),
       Self::MulUnary(lhs, op, rhs) => match (lhs.eval(scopes), rhs.eval(scopes)) {
@@ -67,7 +67,7 @@ impl Evaluate for MulExp {
 }
 
 impl Evaluate for AddExp {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     match self {
       Self::Mul(exp) => exp.eval(scopes),
       Self::AddMul(lhs, op, rhs) => match (lhs.eval(scopes), rhs.eval(scopes)) {
@@ -82,15 +82,15 @@ impl Evaluate for AddExp {
 }
 
 impl Evaluate for RelExp {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     match self {
       Self::Add(exp) => exp.eval(scopes),
       Self::RelAdd(lhs, op, rhs) => match (lhs.eval(scopes), rhs.eval(scopes)) {
         (Some(lhs), Some(rhs)) => Some(match op {
-          RelOp::Lt => (lhs < rhs) as i64,
-          RelOp::Gt => (lhs > rhs) as i64,
-          RelOp::Le => (lhs <= rhs) as i64,
-          RelOp::Ge => (lhs >= rhs) as i64,
+          RelOp::Lt => (lhs < rhs) as i32,
+          RelOp::Gt => (lhs > rhs) as i32,
+          RelOp::Le => (lhs <= rhs) as i32,
+          RelOp::Ge => (lhs >= rhs) as i32,
         }),
         _ => None,
       },
@@ -99,13 +99,13 @@ impl Evaluate for RelExp {
 }
 
 impl Evaluate for EqExp {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     match self {
       Self::Rel(exp) => exp.eval(scopes),
       Self::EqRel(lhs, op, rhs) => match (lhs.eval(scopes), rhs.eval(scopes)) {
         (Some(lhs), Some(rhs)) => Some(match op {
-          EqOp::Eq => (lhs == rhs) as i64,
-          EqOp::Neq => (lhs != rhs) as i64,
+          EqOp::Eq => (lhs == rhs) as i32,
+          EqOp::Neq => (lhs != rhs) as i32,
         }),
         _ => None,
       },
@@ -114,11 +114,11 @@ impl Evaluate for EqExp {
 }
 
 impl Evaluate for LAndExp {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     match self {
       Self::Eq(exp) => exp.eval(scopes),
       Self::LAndEq(lhs, rhs) => match (lhs.eval(scopes), rhs.eval(scopes)) {
-        (Some(lhs), Some(rhs)) => Some((lhs != 0 && rhs != 0) as i64),
+        (Some(lhs), Some(rhs)) => Some((lhs != 0 && rhs != 0) as i32),
         _ => None,
       },
     }
@@ -126,11 +126,11 @@ impl Evaluate for LAndExp {
 }
 
 impl Evaluate for LOrExp {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     match self {
       Self::LAnd(exp) => exp.eval(scopes),
       Self::LOrLAnd(lhs, rhs) => match (lhs.eval(scopes), rhs.eval(scopes)) {
-        (Some(lhs), Some(rhs)) => Some((lhs != 0 || rhs != 0) as i64),
+        (Some(lhs), Some(rhs)) => Some((lhs != 0 || rhs != 0) as i32),
         _ => None,
       },
     }
@@ -138,7 +138,7 @@ impl Evaluate for LOrExp {
 }
 
 impl Evaluate for ConstExp {
-  fn eval(&self, scopes: &Scopes) -> Option<i64> {
+  fn eval(&self, scopes: &Scopes) -> Option<i32> {
     self.exp.eval(scopes)
   }
 }
